+++ src/Data/GameDefinition.cs
@@ -5,11 +5,12 @@ namespace Queens.Data;
 internal readonly record struct GameDefinition
 {
     [JsonPropertyName("sideLength")]
-    public int SideLength { get; init; }
+    public ushort SideLength { get; init; }
 
     [JsonPropertyName("colors")]
     public ColorData[] Colors { get; init; }
 
     [JsonPropertyName("cellColors")]
-    public int[] CellColors { get; init; }
+    public ushort[] CellColors { get; init; }
+
 }
