+++ src/Interfaces/Core/Variables/ICell.cs
@@ -2,7 +2,7 @@ namespace Queens.Interfaces;
 
 internal interface ICell
 {
-    public ushort Id { get; }
+    public int Id { get; }
     public bool Satisfied { get; }
     public IEnumerable<ICell> Edges { get; }
